refactor: externalize core initialization
